requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.6
djangorestframework==3.12.2
pip==20.3.3
psycopg2==2.8.5
pytz==2021.1
setuptools==52.0.0
sqlparse==0.4.1
wheel==0.36.2
wincertstore==0.2

pandas~=1.5.2